# flask_app.zip/app/static/swagger1.json
{
  "openapi": "3.0.0",
  "info": {
      "title": "Shuttle Tracking System",
      "version": "1.0.0",
      "description": "API for user registration",
      "license": {
          "name": "MIT",
          "url": "https://opensource.org/licenses/MIT"
      }
  },
  "paths": {
...
"/duration/distance": {
  "get": {
      "summary": "Calculate total distance and duration",
      "parameters": [
          {
              "name": "current_cord",
              "in": "query",
              "description": "Current coordinates",
              "required": true,
              "schema": {
                  "type": "object",
                  "properties": {
                      "latitude": {
                          "type": "number",
                          "format": "float",
                          "description": "Latitude of current location"
                      },
                      "longitude": {
                          "type": "number",
                          "format": "float",
                          "description": "Longitude of current location"
                      }
                  },
                  "required": ["latitude", "longitude"]
              }
          },
          {
              "name": "destination_cord",
              "in": "query",
              "description": "Destination coordinates",
              "required": true,
              "schema": {
                  "type": "object",
                  "properties": {
                      "latitude": {
                          "type": "number",
                          "format": "float",
                          "description": "Latitude of destination location"
                      },
                      "longitude": {
                          "type": "number",
                          "format": "float",
                          "description": "Longitude of destination location"
                      }
                  },
                  "required": ["latitude", "longitude"]
              }
          },
          {
              "name": "intermediate_cord",
              "in": "query",
              "description": "Intermediate coordinates",
              "required": true,
              "schema": {
                  "type": "array",
                  "items": {
                      "type": "object",
                      "properties": {
                          "latitude": {
                              "type": "number",
                              "format": "float",
                              "description": "Latitude of intermediate location"
                          },
                          "longitude": {
                              "type": "number",
                              "format": "float",
                              "description": "Longitude of intermediate location"
                          }
                      },
                      "required": ["latitude", "longitude"]
                  }
              }
          }
      ],
      "responses": {
          "200": {
              "description": "Distance and duration calculated successfully",
              "content": {
                  "application/json": {
                      "example": {
                          "distance": 10.5,
                          "duration": 15.3
                      }
                  }
              }
          },
          "400": {
              "description": "Bad Request",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "Invalid request data"
                      }
                  }
              }
          },
          "500": {
              "description": "Internal Server Error",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "An error occurred during calculation. Please try again."
                      }
                  }
              }
          }
      }
  },
  "post": {
      "summary": "Calculate total distance and duration",
      "requestBody": {
          "content": {
              "application/json": {
                  "schema": {
                      "type": "object",
                      "properties": {
                          "current_cord": {
                              "type": "object",
                              "description": "Current coordinates",
                              "properties": {
                                  "latitude": {
                                      "type": "number",
                                      "format": "float",
                                      "description": "Latitude of current location"
                                  },
                                  "longitude": {
                                      "type": "number",
                                      "format": "float",
                                      "description": "Longitude of current location"
                                  }
                              },
                              "required": ["latitude", "longitude"]
                          },
                          "destination_cord": {
                              "type": "object",
                              "description": "Destination coordinates",
                              "properties": {
                                  "latitude": {
                                      "type": "number",
                                      "format": "float",
                                      "description": "Latitude of destination location"
                                  },
                                  "longitude": {
                                      "type": "number",
                                      "format": "float",
                                      "description": "Longitude of destination location"
                                  }
                              },
                              "required": ["latitude", "longitude"]
                          },
                          "intermediate_cord": {
                              "type": "array",
                              "description": "Intermediate coordinates",
                              "items": {
                                  "type": "object",
                                  "properties": {
                                      "latitude": {
                                          "type": "number",
                                          "format": "float",
                                          "description": "Latitude of intermediate location"
                                      },
                                      "longitude": {
                                          "type": "number",
                                          "format": "float",
                                          "description": "Longitude of intermediate location"
                                      }
                                  },
                                  "required": ["latitude", "longitude"]
                              }
                          }
                      },
                      "required": ["current_cord", "destination_cord", "intermediate_cord"]
                  }
              }
          }
      },
      "responses": {
          "200": {
              "description": "Distance and duration calculated successfully",
              "content": {
                  "application/json": {
                      "example": {
                          "distance": 10.5,
                          "duration": 15.3
                      }
                  }
              }
          },
          "400": {
              "description": "Bad Request",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "Invalid request data"
                      }
                  }
              }
          },
          "500": {
              "description": "Internal Server Error",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "An error occurred during calculation. Please try again."
                      }
                  }
              }
          }
      }
  }
},
"/current_cord": {
  "get": {
      "summary": "Get current coordinates",
      "parameters": [
          {
              "name": "device_id",
              "in": "query",
              "description": "ID of the device",
              "required": true,
              "schema": {
                  "type": "string",
                  "description": "The device ID"
              }
          }
      ],
      "responses": {
          "200": {
              "description": "Current coordinates retrieved successfully",
              "content": {
                  "application/json": {
                      "example": {
                          "latitude": 12.34,
                          "longitude": 56.78
                      }
                  }
              }
          },
          "400": {
              "description": "Bad Request",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "Invalid request data"
                      }
                  }
              }
          },
          "500": {
              "description": "Internal Server Error",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "An error occurred. Please try again."
                      }
                  }
              }
          }
      }
  },
  "post": {
      "summary": "Get current coordinates",
      "requestBody": {
          "content": {
              "application/x-www-form-urlencoded": {
                  "schema": {
                      "type": "object",
                      "properties": {
                          "device_id": {
                              "type": "string",
                              "description": "ID of the device"
                          }
                      },
                      "required": ["device_id"]
                  }
              }
          }
      },
      "responses": {
          "200": {
              "description": "Current coordinates retrieved successfully",
              "content": {
                  "application/json": {
                      "example": {
                          "latitude": 12.34,
                          "longitude": 56.78
                      }
                  }
              }
          },
          "400": {
              "description": "Bad Request",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "Invalid request data"
                      }
                  }
              }
          },
          "500": {
              "description": "Internal Server Error",
              "content": {
                  "application/json": {
                      "example": {
                          "error": "An error occurred. Please try again."
                      }
                  }
              }
          }
      }
  }
}
  }
}